Added presenting alert functionality.
